Working with CloudFormation Nested Stacks
About this lab

In this hands-on lab, we will create a nested stack from the templates provided. We will use the two templates to create a nested stack that will implement a child template, which will then create an S3 bucket with a policy of no retain. This child template can be reused repeatedly whenever an S3 bucket of this type is needed.

Learning objectives

Create an S3 Bucket and Upload CloudFormation Templates

Create a Nested Stack with a Single Child Stack

Create a Nested Stack with a Multiple Child Stacks
Upload the index.html file and verify the page loads